@@ -1,43 +1,43 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.13'
+ruby '2.1.2'
 
-group :development do
-  gem 'sqlite3'
+gem 'rails', '~> 4.0'
+
+gem 'pg'
+gem 'thin'
+
+group :development, :test do
+  gem 'spring'
+  gem 'better_errors'
+  gem 'binding_of_caller'
 end
 
-group :production do
-  # gem 'pg'
-end
-
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'therubyracer', :platforms => :ruby
-  gem 'uglifier', '>= 1.0.3'
-  gem "compass-rails"
-  gem "compass-rgbapng", require: "rgbapng"
-  gem "bootstrap-sass"
-end
+gem 'sass-rails', "4.0.3"
+gem 'coffee-rails'
+gem 'therubyracer', :platforms => :ruby
+gem 'uglifier', '>= 1.0.3'
+gem 'compass-rails', github: 'Compass/compass-rails'
+gem "compass-rgbapng", require: "rgbapng"
+gem "bootstrap-sass", "3.2.0.0"
 
 gem 'jquery-rails'
 
 # gem 'rails_admin_glyph_theme', git: 'git://github.com/vala/rails_admin_glyph_theme.git'
 
-# gem 'rails_admin'
-gem 'rails_admin', git: "git://github.com/vala/rails_admin.git"
+gem 'rails_admin'
+# gem 'rails_admin', github: "vala/rails_admin"
 # gem 'rails_admin', path: "/Users/vala/Documents/development/ruby/gems/rails_admin"
 
 gem 'paperclip'
 gem 'devise'
 gem 'simple_form'
 
-
 # gem 'restful_sync', git: "git://github.com/vala/restful_sync.git"
 # gem 'restful_sync', path: "/Users/damien/glyph/gems/restful_sync"
-# gem 'glysellin', path: "/Users/vala/Documents/development/ruby/gems/glysellin"
-gem 'glysellin', git: "git://github.com/Glysellin/glysellin.git"
-gem 'glysellin_associated_products', git: "git://github.com/Glysellin/glysellin_associated_products.git"
+
+gem 'glysellin', github: "glysellin/glysellin"
+# gem 'glysellin', path: "~/gems/glysellin"
+
+# gem 'glysellin_associated_products', github: "glysellin/glysellin_associated_products"
 # gem 'glysellin_associated_products', path: "/Users/damien/glyph/gems/glysellin/glysellin_associated_products"
