@@ -12,7 +12,7 @@ end
 
 module GlysellinTest
   class Application < Rails::Application
-    
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
@@ -28,7 +28,7 @@ module GlysellinTest
     # config.active_record.observers = :cacher, :garbage_collector, :forum_observer
 
     # config.active_record.observers = RestfulSync::ApiObserver
-    
+
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
@@ -52,12 +52,6 @@ module GlysellinTest
     # config.active_record.schema_format = :sql
 
     config.assets.initialize_on_precompile = false
-
-    # Enforce whitelist mode for mass assignment.
-    # This will create an empty whitelist of attributes available for mass-assignment for all models
-    # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
-    # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = true
 
     # Enable the asset pipeline
     config.assets.enabled = true
